Split sample and dataset stuff in IUT and MIP.
